Simplify PDF extraction implementation
