+++ Sources/Nativ/Features/Chat/ChatAttachmentValidator.swift
@@ -1,20 +1,33 @@
 import Foundation
 
+enum ChatAttachmentValidation: Equatable, Sendable {
+    case processing(message: String)
+    case ready(extractedCharacterCount: Int?)
+    case warning(message: String, extractedCharacterCount: Int?)
+    case blocked(message: String)
+
+    var preventsSending: Bool {
+        switch self {
+        case .processing, .blocked:
+            true
+        case .ready, .warning:
+            false
+        }
+    }
+
+    var extractedCharacterCount: Int? {
+        switch self {
+        case .ready(let count), .warning(_, let count):
+            count
+        case .processing, .blocked:
+            nil
+        }
+    }
+}
+
 actor ChatAttachmentValidator {
     private let extractionCache: ChatDocumentExtractionCache
     private let maximumCharactersPerDocument: Int
-
-    init(
-        extractor: any DocumentTextExtracting = PDFDocumentTextExtractor(),
-        maximumCharactersPerDocument: Int = ChatDocumentContextBuilder.defaultMaximumCharactersPerDocument
-    ) {
-        precondition(maximumCharactersPerDocument > 0)
-        self.extractionCache = ChatDocumentExtractionCache(
-            extractor: extractor,
-            maximumCharactersPerDocument: maximumCharactersPerDocument
-        )
-        self.maximumCharactersPerDocument = maximumCharactersPerDocument
-    }
 
     init(extractionCache: ChatDocumentExtractionCache) {
         self.extractionCache = extractionCache
