+++ project.yml
@@ -216,7 +216,6 @@ targets:
       - path: Sources/Nativ/Features/Documents
       - path: Sources/Nativ/Features/Chat/ChatSessionStore.swift
       - path: Sources/Nativ/Features/Chat/ChatAttachmentKind.swift
-      - path: Sources/Nativ/Features/Chat/ChatAttachmentValidation.swift
       - path: Sources/Nativ/Features/Chat/ChatAttachmentValidator.swift
       - path: Sources/Nativ/Features/Chat/ChatDocumentExtractionCache.swift
       - path: Sources/Nativ/Features/Chat/ChatDocumentContextBuilder.swift
